Adding and creating initial hdf5 gas database
This hdf5 database shall have all needed data for each gas. Hdf5 was used for it's SWMR (Single Writer Multiple Readers) properties. This property will be used when implementing a standard and fluid gas function for all used gases. 

Furthermore, this python script shall have a user interface that allows the user to setup and add a gas into the database.